Completely revamped the Callback system [API Change]
Callbacks now need to be assigned to a member variable to keep it alive.
There is some backwards compatibility, but it's ill advised as it will
leak memory if used incorrectly.

You now must use Callback<Type_t>.Create(Function); in OnEnable.
And make sure that you are holding a reference to it or the GC will
clean it up before you even get to use it!